# contracts/openapi.yaml
openapi: 3.0.3
info:
  title: AI ReqEval API
  version: 0.1.0
servers:
  - url: http://127.0.0.1:8000
paths:
...
  /api/v1/skills:
    get:
      summary: List fixed skills
      tags: [skills]
      parameters:
        - name: stage_id
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Skill list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SkillListResponse"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/skills/{skill_id}/invoke:
    post:
      summary: Invoke fixed skill
      tags: [skills]
      parameters:
        - $ref: "#/components/parameters/SkillId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvokeSkillRequest"
      responses:
        "200":
          description: Skill invocation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SkillInvokeResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
...
components:
  parameters:
...
    SkillId:
      name: skill_id
      in: path
      required: true
      schema:
        type: string
...
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: Request conflicts with current resource state
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        detail:
          oneOf:
            - type: string
            - type: array
              items:
                type: object
      required: [detail]
...
    SkillResponse:
      type: object
      properties:
        name: { type: string }
        stage_suffix: { type: string }
        description: { type: string }
        input_schema:
          type: object
          additionalProperties:
            type: string
        output_schema:
          type: object
          additionalProperties:
            type: string
        allowed_tools:
          type: array
          items: { type: string }
        allowed_subagents:
          type: array
          items: { type: string }
        auto_run_condition: { type: string }
        visibility: { type: string }
      required: [name, stage_suffix, description, input_schema, output_schema, allowed_tools, allowed_subagents, auto_run_condition, visibility]
    SkillListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/SkillResponse"
      required: [items]
    InvokeSkillRequest:
      type: object
      properties:
        project_id: { type: string }
        stage_id: { type: string }
        run_id:
          type: string
          nullable: true
        goal:
          type: string
          minLength: 1
      required: [project_id, stage_id, goal]
    SkillInvokeResponse:
      type: object
      properties:
        skill_name: { type: string }
        stage_id: { type: string }
        project_id: { type: string }
        summary: { type: string }
        tool_names:
          type: array
          items: { type: string }
        subagent_names:
          type: array
          items: { type: string }
        enabled_tools:
          type: array
          items: { type: string }
        enabled_subagents:
          type: array
          items: { type: string }
        config_version_id:
          type: string
          nullable: true
        goal: { type: string }
        stage_result_id: { type: string }
        report_id:
          type: string
          nullable: true
        tool_results:
          type: array
          items:
            type: object
            additionalProperties: true
        warnings:
          type: array
          items:
            type: object
            additionalProperties: true
      required: [skill_name, stage_id, project_id, summary, tool_names, subagent_names, goal, stage_result_id, tool_results]
...